First pass at adding an abstract way to handle Qt UI parenting.
Adds a public function to the maya engine that accepts a Qt Widget class.
The purpose of this function is to give abstract access to parenting
Qt Widgets with various host applications and have it be available
in all engines.  This alleviates having to have host specific code
in each app that offers a UI.