#29765 Added support for toolkit panels.
It's now possible to register panels in maya using the new panel
interface methods that are being introduced in core. This change
implements the show_panel() method, making it possible to start a new
panel. The panels are 'singletons', meaning that a UI can only exist
once. Clicking it again on the menu will result in the existing panel
being brought forward.

Closes #7