#37277 Support for site context
This removes a previous check in the maya engine to ensure that at least a project context existed at startup. This check seems unnecessary and prevents the maya engine from starting up in a site-mode plugin setup. I think it was there to ensure that a maya project was always set by the engine - but we loosened that restriction a long time ago and the maya engine often runs with no project set. So it should be ok to just drop this check.

Closes #13