#34893: Run toolkit manager bootstrap logic in a background thread.
When logging in the plug-in to Shotgun, bootstrap the toolkit asynchronously in a background thread, followed by launching the engine synchronously in the main application thread. This allows Maya to continue its execution and remain responsive when bootstrapping the toolkit involves downloading files and installing apps from the toolkit app store.

Maya shows a progress bar in the Help Line while the engine bootstraps in background (#37234).

The plug-in now works with older versions of Maya (#38089).