[minor] More logging during engine launch (#40)
This adds support for the new bootstrap.pre_engine_start_callback that was recently introduced. Previously, plugin logging would be turned off before bootstrap started, there would then be a gap where logging would go to the backend loggers but not to the UI and the UI logging would kick in again as the engine starts up. This tweak narrows the gap where no UI logging exists.