#41519 Cleanup of maya plugin logging (#47)
I have been investigating stability issues where maya would hang on windows maya startup. It basically boils down to the synchronous log dispatch from the plugin. I took a look at the code and realised that there was a bunch of custom logging going on which I have removed, just to reduce the future support and engineering burden. There is a downside to this - the maya plugin only logs to the backend log file during startup. Once the engine kicks in, the in-console logging starts. For any troubleshooting at this stage, the log files contain all the info. Also polished the ux and messaging around startup.