SG-9012: Ensures that progress updates are always run in the main thread during bootstrap. (#67)
Updating Maya's progress bar is NOT thread safe, so we need to ensure that all updates of progress there are processed in the main thread. The callback function registered with the bootstrapper is now only responsible for setting state on an object, and a QTimer running on the main thread updates the progress bar when necessary. This addresses a freeze/crash bug on launch of Maya when tk-core v0.18.153 is in use.